// src/typings/Company.d.ts
export interface Company {
    dossier_number: string;
    establishment_number: string;
    name: string;
    match_type: string;
    establishment_city: string;
    establishment_street: string;
    correspondence_city: string;
    correspondence_street: string;
    indication_economically_active: boolean;
}